mypy updates
